Refactor emotion theme code, replace promise with async await, fix route issue
