+++ client/my-sites/plugins/index.js
@@ -104,7 +104,7 @@ export default function () {
 		);
 
 		if ( config.isEnabled( 'marketplace-yoast' ) ) {
-			page( '/plugins/domain/:site?', siteSelection, renderDomainsPage, makeLayout, clientRender );
+			page( '/plugins/domain/:site', siteSelection, renderDomainsPage, makeLayout, clientRender );
 		}
 		page(
 			'/plugins/:pluginFilter(active|inactive|updates)/:site_id?',
